# SolucionHyJ_20191025 (Fase 1)/WS_HyJ/Controllers/AccountController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Identity;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Configuration;
using WS_HyJ.Helpers;
using WS_HyJ.Models.Identity;
using WS_HyJ.Models.Requests;
using WS_HyJ.Models.Responses;
using WS_HyJ.Seguridad;

namespace WS_HyJ.Controllers
{
    [Route("auth")]
    [ApiController]
    [ApiVersion("1.0")]
    public class AccountController : ControllerBase
    {
        private readonly UserManager<ApplicationUser> _userManager;
        private readonly SignInManager<ApplicationUser> _signInManager;
        private readonly RoleManager<ApplicationRole> _roleManager;
        private readonly IConfiguration _configuration;
        private readonly IAuthenticateService _authService;

        public AccountController(UserManager<ApplicationUser> userManager, RoleManager<ApplicationRole> roleManager, SignInManager<ApplicationUser> signInManager, IConfiguration configuration, IAuthenticateService authService)
        {
            _roleManager = roleManager;
            _userManager = userManager;
            _signInManager = signInManager;
            _configuration = configuration;
            _authService = authService;
        }

        /// <summary>
        /// Método de logueo para consumir los servicios internos.
        /// </summary>
        /// <param name="request">Modelo de logueo</param>
        /// <returns></returns>
        [AllowAnonymous]
        [HttpPost, Route("login")]
        public IActionResult RequestToken([FromForm] LoginDTO request)
        {
            if (!ModelState.IsValid)
            {
                return BadRequest(ModelState);
            }

            if (_authService.IsAuthenticated(request, out string token))
            {
                return Ok(token);
            }

            return BadRequest("Credenciales inválidas");
        }

        [HttpPost, Route("register")]
        [AllowAnonymous]
        public async Task<IActionResult> Register([FromBody]RegisterDTO model)
        {
            if (ModelState.IsValid)
            {
                var user = new ApplicationUser { Name = model.Name, LastName = model.LastName, UserName = model.Email, Email = model.Email };
                var result = await _userManager.CreateAsync(user, model.Password);
                if (result.Succeeded)
                {                    
                    var rootData = new SignUpResponse(user.UserName, user.Email);
                    return Created("api/account/register", rootData);
                }
                return Ok(string.Join(",", result.Errors?.Select(error => error.Description)));
            }
            string errorMessage = string.Join(", ", ModelState.Values.SelectMany(x => x.Errors).Select(x => x.ErrorMessage));
            return BadRequest(errorMessage ?? "Bad Request");
        }
    }
}