# SolucionHyJ_20191025 (Fase 1)/WS_HyJ/wwwroot/swagger.yaml
swagger: "2.0"
info:
  version: 1.0
  title: REST API - HyJ
  description: Lista de servicios API para la solucion de HyJ.
  termsOfService: ''
  contact:
    name: Diana Rolando
    url: ''
    email: drolando@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /DEV/api/v1/auth/login:
    post:
      tags:
      - Account
      summary: Método de logueo para consumir los servicios internos.
      operationId: RequestToken
      consumes:
      - multipart/form-data
      produces: []
      parameters:
      - name: Username
        in: formData
        required: true
        type: string
      - name: Password
        in: formData
        required: true
        type: string
      responses:
        200:
          description: Success
...
  /DEV/api/v1/Brand/{id}:
    get:
      tags:
      - Brand
      operationId: GetBrand
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/BrandEntity'
    put:
      tags:
      - Brand
      operationId: Put
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: model
        in: body
        schema:
          $ref: '#/definitions/BrandEntity'
      responses:
        200:
          description: Success
    delete:
      tags:
      - Brand
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
...
  /DEV/api/v1/Kardex/{id}:
    get:
      tags:
      - Kardex
      operationId: GetPI
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/KardexEntity'
    put:
      tags:
      - Kardex
      operationId: Put
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: model
        in: body
        schema:
          $ref: '#/definitions/KardexRequest'
      responses:
        200:
          description: Success
    delete:
      tags:
      - Kardex
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
  /DEV/api/v1/Kardex/GetStock/{idProvider}:
    get:
      tags:
      - Kardex
      operationId: GetStock
      consumes: []
      produces:
      - application/json
      parameters:
      - name: idProvider
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/ProductsByProvider'
...
  /DEV/api/v1/Option/UploadImage:
    post:
      tags:
      - Option
      summary: Uplaods an image to the server.
      operationId: UploadImage
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - name: file
        in: formData
        description: ''
        type: file
      responses:
        200:
          description: Success
...
definitions:
...
  BrandEntity:
    description: Entidad de marcas
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      nombre:
        type: string
  KardexEntity:
    description: Kardex
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      idProducto:
        type: string
      minimumAmount:
        format: int32
        type: integer
      currentAmount:
        format: int32
        type: integer
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
  KardexRequest:
    required:
    - idProducto
    type: object
    properties:
      idProducto:
        type: string
      minimumAmount:
        format: int32
        type: integer
      currentAmount:
        format: int32
        type: integer
  ProductsByProvider:
    type: object
    properties:
      product:
        $ref: '#/definitions/ProductEntity'
      minimumAmount:
        format: int32
        type: integer
      currentAmount:
        format: int32
        type: integer
  ProductEntity:
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      numero:
        format: int32
        type: integer
      codigo:
        description: Correlativo (P-0001)
        type: string
      nombre:
        type: string
      precioUnitario:
        format: double
        type: number
      moneda:
        enum:
        - Soles
        - Dolares
        type: string
      um:
        enum:
        - Unidad
        - Miligramo
        - Gramo
        - Kilogramo
        - Litro
        type: string
      peso:
        format: double
        type: number
      tipo:
        type: string
      descripcion:
        type: string
      idProveedor:
        type: string
      marca:
        $ref: '#/definitions/BrandEntity'
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
...
securityDefinitions:
  Bearer:
    name: Authorization
    in: header
    type: apiKey
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
security:
- Bearer: []